+++ I Bike CPH/CykelPlanen/Classes/SMTrain.h
@@ -6,17 +6,21 @@
 //  Copyright (c) 2013 City of Copenhagen. All rights reserved.
 //
 
-#import <Foundation/Foundation.h>
 #import "SMArrivalInformation.h"
+#import <Foundation/Foundation.h>
 
 /**
- * Train. Has arrival information for stations. FIXME: arrival inforamtion is exposed as mutable/nilable array. Used in SMRouteInfoViewController, SMTransportation.
+ * Train. Has arrival information for stations. FIXME: arrival inforamtion is exposed as mutable/nilable array. Used in SMRouteInfoViewController,
+ * SMTransportation.
  */
 @interface SMTrain : NSObject
 
-@property(nonatomic, strong) NSMutableArray* arrivalInformation;
+@property(nonatomic, strong) NSMutableArray *arrivalInformation;
 
--(SMArrivalInformation*)informationForStation:(SMStationInfo*)station;
--(NSArray*)routeTimestampsForSourceStation:(SMStationInfo*)sourceSt destinationStation:(SMStationInfo*)destinationSt forDay:(int)dayIndex time:(SMTime*)time;
+- (SMArrivalInformation *)informationForStation:(SMStationInfo *)station;
+- (NSArray *)routeTimestampsForSourceStation:(SMStationInfo *)sourceSt
+                          destinationStation:(SMStationInfo *)destinationSt
+                                      forDay:(NSInteger)dayIndex
+                                        time:(SMTime *)time;
 
 @end
